Move ActiveRecord::ConnectionAdapters::ConnectionPool::Repaer to separate file
