+++ activerecord/lib/active_record/connection_adapters.rb
@@ -25,16 +25,13 @@ module ActiveRecord
       autoload :ReferenceDefinition
     end
 
-    autoload_at "active_record/connection_adapters/abstract/connection_handler" do
-      autoload :ConnectionHandler
-    end
-
     autoload_under "abstract" do
       autoload :SchemaStatements
       autoload :DatabaseStatements
       autoload :DatabaseLimits
       autoload :Quoting
       autoload :ConnectionPool
+      autoload :ConnectionHandler
       autoload :QueryCache
       autoload :Savepoints
     end
